akbank-rag-chatbot/requirements.txt
# === Core ===
python-dotenv==1.0.1
pandas==2.2.2
numpy==1.26.4

# === API & Web Framework ===
fastapi==0.110.0
uvicorn==0.29.0

# === LangChain & Gemini ===
langchain==0.1.20
langchain-google-genai==1.0.3

# === Vector Database ===
chromadb==0.4.24

# === Utility ===
tqdm==4.66.4
streamlit==1.39.0